Update setup-pi4.sh
make radar display install optional

+++ setup-pi4.sh
@@ -46,11 +46,17 @@ cd kalibrate-rtl
 make && make install
 
 # install stratux-radar-display
-cd /root
-rm -rf /root/stratux-radar-display
-apt install libatlas-base-dev zlib1g-dev libfreetype6-dev liblcms2-dev libopenjp2-7 libtiff5 python3-pip python3-pil espeak-ng espeak-ng-data libespeak-ng-dev libbluetooth-dev -y
-pip3 install luma.oled websockets py-espeak-ng pybluez pydbus numpy
-git clone https://github.com/TomBric/stratux-radar-display.git
+echo
+read -t 1 -n 10000 discard
+read -p "Install Radar Display? [y/n]"
+echo
+if [[ $REPLY =~ ^[Yy]$ ]]; then
+  cd /root
+  rm -rf /root/stratux-radar-display
+  apt install libatlas-base-dev zlib1g-dev libfreetype6-dev liblcms2-dev libopenjp2-7 libtiff5 python3-pip python3-pil espeak-ng espeak-ng-data libespeak-ng-dev libbluetooth-dev -y
+  pip3 install luma.oled websockets py-espeak-ng pybluez pydbus numpy
+  git clone https://github.com/TomBric/stratux-radar-display.git
+fi
 
 # clone stratux
 cd /root
